Refactor damage radius to float type and update related methods; add PoolManager script for pixel object pooling

+++ Assets/_Scripts/SOs/LevelConfig.cs
@@ -11,7 +11,7 @@ public class LevelConfig : ScriptableObject
     public List<BlockData> blocksToSpawn = new List<BlockData>(); 
 
     [Header("Tap damage radius")]
-    public int damageRadius = 5;
+    public float damageRadius = 5;
     public int maxTapDamage = 3;
     public int minTapDamage = 1;
 }
